+++ src/main/java/com/domain/service/ProductService.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public interface ProductService {
     List<ProductDto> getAllProducts();
-    ProductDto findProductByProductId(int productId);
+    ProductDto findProductByProductId(Long productId);
     Long registerProduct(ProductDto productDto);
-    ProductDto updateProductDetail(int productId, ProductDto productDto);
-    int removeProduct(int productId);
+    ProductDto updateProductDetail(Long productId, ProductDto productDto);
+    int removeProduct(Long productId);
     List<ProductDto> findProductsBySearchingKeyword(String keyword, int price, boolean sale);
 
 }
